localml/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appnope==0.1.0            # via ipykernel, ipython
argon2-cffi==20.1.0       # via notebook
arrow==0.17.0             # via -r requirements.in
async-generator==1.10     # via nbclient
attrs==20.3.0             # via jsonschema
azure-core==1.9.0         # via azure-storage-blob
azure-storage-blob==12.6.0  # via -r requirements.in
backcall==0.2.0           # via ipython
bleach==3.2.1             # via nbconvert
certifi==2020.11.8        # via msrest, requests
cffi==1.14.4              # via argon2-cffi, cryptography
chardet==3.0.4            # via requests
cryptography==3.3.1       # via azure-storage-blob
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, networkx
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
idna==2.10                # via requests
imageio==2.9.0            # via scikit-image
ipykernel==5.3.4          # via -r requirements.in, ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.19.0           # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
isodate==0.6.0            # via msrest
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via nbconvert, notebook
joblib==0.17.0            # via scikit-learn
jsonschema==3.2.0         # via nbformat
jupyter-client==6.1.7     # via ipykernel, jupyter-console, nbclient, notebook, qtconsole
jupyter-console==6.2.0    # via jupyter
jupyter-core==4.7.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r requirements.in
jupyterlab-pygments==0.1.2  # via nbconvert
kiwisolver==1.3.1         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.3         # via -r requirements.in, scikit-image, seaborn
mistune==0.8.4            # via nbconvert
msrest==0.6.19            # via azure-storage-blob
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via jupyter, notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, notebook
nest-asyncio==1.4.3       # via nbclient
networkx==2.5             # via scikit-image
notebook==6.1.5           # via jupyter, widgetsnbextension
numpy==1.19.4             # via -r requirements.in, imageio, matplotlib, pandas, pywavelets, scikit-image, scikit-learn, scipy, seaborn, tifffile
oauthlib==3.1.0           # via requests-oauthlib
packaging==20.4           # via bleach
pandas==1.1.4             # via -r requirements.in, seaborn
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==8.0.1             # via -r requirements.in, imageio, matplotlib, scikit-image
prometheus-client==0.9.0  # via notebook
prompt-toolkit==3.0.8     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
pycparser==2.20           # via cffi
pygments==2.7.2           # via ipython, jupyter-console, jupyterlab-pygments, nbconvert, qtconsole
pyparsing==2.4.7          # via matplotlib, packaging
pyrsistent==0.17.3        # via jsonschema
python-dateutil==2.8.1    # via arrow, jupyter-client, matplotlib, pandas
pytz==2020.4              # via pandas
pywavelets==1.1.1         # via scikit-image
pyzmq==20.0.0             # via jupyter-client, notebook, qtconsole
qtconsole==5.0.1          # via jupyter
qtpy==1.9.0               # via qtconsole
requests-oauthlib==1.3.0  # via msrest
requests==2.25.0          # via -r requirements.in, azure-core, msrest, requests-oauthlib
scikit-image==0.17.2      # via -r requirements.in
scikit-learn==0.23.2      # via -r requirements.in
scipy==1.5.4              # via -r requirements.in, scikit-image, scikit-learn, seaborn
seaborn==0.11.0           # via -r requirements.in
send2trash==1.5.0         # via notebook
six==1.15.0               # via argon2-cffi, azure-core, bleach, cryptography, cycler, isodate, jsonschema, packaging, python-dateutil
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
threadpoolctl==2.1.0      # via scikit-learn
tifffile==2020.11.18      # via scikit-image
tornado==6.1              # via ipykernel, jupyter-client, notebook, terminado
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, qtconsole
urllib3==1.26.2           # via requests
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets

# The following packages are considered to be unsafe in a requirements file:
# setuptools
